test: cover validated CLI hint in CI lane
